remove request base namespace (move up a directory)
