+++ DragonFruit.Six.Api/Requests/AccountActivityRequest.cs
@@ -6,7 +6,6 @@ using System.Linq;
 using DragonFruit.Common.Data;
 using DragonFruit.Common.Data.Parameters;
 using DragonFruit.Six.Api.Entities;
-using DragonFruit.Six.Api.Requests.Base;
 using DragonFruit.Six.Api.Utils;
 
 namespace DragonFruit.Six.Api.Requests
