+++ DragonFruit.Six.Api/Requests/SeasonStatsRequest.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using DragonFruit.Common.Data;
 using DragonFruit.Common.Data.Parameters;
 using DragonFruit.Six.Api.Entities;
-using DragonFruit.Six.Api.Requests.Base;
 
 #pragma warning disable 628
